@@ -1,6 +1,6 @@
 {
   "name": "angular-starter",
-  "version": "8.2.14",
+  "version": "9.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve --port 4200",
@@ -8,42 +8,43 @@
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "compile:server": "webpack --config webpack.server.config.js --progress --colors",
-    "serve:ssr": "node dist/server",
-    "build:ssr": "npm run build:client-and-server-bundles && npm run compile:server",
-    "build:client-and-server-bundles": "ng build --prod && ng run angular-starter:server:production --bundleDependencies all"
+    "dev:ssr": "ng run angular-starter:serve-ssr",
+    "serve:ssr": "node dist/angular-starter/server/main.js",
+    "build:ssr": "ng build --prod && ng run angular-starter:server:production",
+    "prerender": "ng run angular-starter:prerender"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.2.14",
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "8.2.14",
-    "@angular/core": "8.2.14",
-    "@angular/forms": "8.2.14",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/platform-server": "8.2.14",
-    "@angular/pwa": "0.803.22",
-    "@angular/router": "8.2.14",
-    "@angular/service-worker": "8.2.14",
-    "@fortawesome/fontawesome-free": "5.12.0",
-    "@nguniversal/express-engine": "8.2.6",
-    "@nguniversal/module-map-ngfactory-loader": "v8.2.6",
+    "@angular/animations": "9.0.0",
+    "@angular/common": "9.0.0",
+    "@angular/compiler": "9.0.0",
+    "@angular/core": "9.0.0",
+    "@angular/forms": "9.0.0",
+    "@angular/platform-browser": "9.0.0",
+    "@angular/platform-browser-dynamic": "9.0.0",
+    "@angular/platform-server": "9.0.0",
+    "@angular/router": "9.0.0",
+    "@angular/service-worker": "9.0.0",
+    "@fortawesome/fontawesome-free": "5.12.1",
+    "@nguniversal/express-engine": "9.0.0",
     "bootstrap": "4.4.1",
+    "chart.js": "2.9.3",
     "express": "4.17.1",
     "jquery": "3.4.1",
+    "ng2-charts": "2.3.0",
     "rxjs": "6.5.4",
     "tslib": "1.10.0",
     "zone.js": "0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.803.22",
-    "@angular/cli": "8.3.22",
-    "@angular/compiler-cli": "8.2.14",
-    "@angular/language-service": "8.2.14",
+    "@angular-devkit/build-angular": "0.900.1",
+    "@angular/cli": "9.0.1",
+    "@angular/compiler-cli": "9.0.0",
+    "@angular/language-service": "9.0.0",
+    "@nguniversal/builders": "9.0.0",
     "@types/express": "4.17.2",
-    "@types/node": "13.1.6",
-    "@types/jasmine": "3.5.0",
+    "@types/node": "13.7.0",
+    "@types/jasmine": "3.5.3",
     "@types/jasminewd2": "2.0.8",
     "codelyzer": "5.2.1",
     "jasmine-core": "3.5.0",
@@ -51,13 +52,11 @@
     "karma": "4.4.1",
     "karma-chrome-launcher": "3.1.0",
     "karma-coverage-istanbul-reporter": "2.1.1",
-    "karma-jasmine": "3.1.0",
-    "karma-jasmine-html-reporter": "1.5.1",
-    "protractor": "5.4.2",
-    "ts-loader": "6.2.1",
-    "ts-node": "8.6.1",
-    "tslint": "5.20.1",
-    "typescript": "3.5.3",
-    "webpack-cli": "3.3.10"
+    "karma-jasmine": "3.1.1",
+    "karma-jasmine-html-reporter": "1.5.2",
+    "protractor": "5.4.3",
+    "ts-node": "8.6.2",
+    "tslint": "6.0.0",
+    "typescript": "3.7.5"
   }
 }
