@@ -1,5 +1,5 @@
+import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 
 import { AppComponent } from './app.component';
 import { HomeComponent } from './modules/general/home/home.component';
@@ -14,9 +14,9 @@ import { HttpClientModule } from '@angular/common/http';
     NotFoundComponent,
   ],
   imports: [
-    CommonModule,
+    BrowserModule,
     AppRoutingModule,
-    HttpClientModule,
+    HttpClientModule
   ],
   providers: [],
   bootstrap: [AppComponent]
