@@ -5,36 +5,54 @@ import { HomeComponent } from './modules/general/home/home.component';
 import { NotFoundComponent } from './modules/general/not-found/not-found.component';
 
 const routes: Routes = [
+  { path: '', component: HomeComponent, },
   {
     path: 'httpclient',
-    loadChildren: () => import('./modules/application/items/items.module').then(mod => mod.ItemsModule)
+    loadChildren: () => import('./modules/application/items/items.module')
+      .then(mod => mod.ItemsModule)
+  },
+  {
+    path: 'template-driven-forms',
+    loadChildren: () => import('./modules/application/template-driven-forms/template-driven-forms.module')
+      .then(mod => mod.TemplateDrivenFormsModule)
+  },
+  {
+    path: 'chartjs',
+    loadChildren: () => import('./modules/application/chartjs/chartjs.module')
+      .then(mod => mod.ChartjsModule)
   },
   {
     path: 'services',
-    loadChildren: () => import('./modules/application/services/services.module').then(mod => mod.ServicesModule)
+    loadChildren: () => import('./modules/application/services/services.module')
+      .then(mod => mod.ServicesModule)
   },
   {
     path: 'components',
-    loadChildren: () => import('./modules/application/components/components.module').then(mod => mod.ComponentsModule)
+    loadChildren: () => import('./modules/application/components/components.module')
+      .then(mod => mod.ComponentsModule)
   },
   {
     path: 'about',
-    loadChildren: () => import('./modules/general/about/about.module').then(mod => mod.AboutModule)
+    loadChildren: () => import('./modules/general/about/about.module')
+      .then(mod => mod.AboutModule)
   },
   {
     path: 'contact',
-    loadChildren: () => import('./modules/general/contact/contact.module').then(mod => mod.ContactModule)
+    loadChildren: () => import('./modules/general/contact/contact.module')
+      .then(mod => mod.ContactModule)
   },
   {
     path: 'signin',
-    loadChildren: () => import('./modules/general/signin/signin.module').then(mod => mod.SigninModule)
+    loadChildren: () => import('./modules/general/signin/signin.module')
+      .then(mod => mod.SigninModule)
   },
-  { path: '', component: HomeComponent, },
   { path: '**', component: NotFoundComponent }
 ];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes)],
+  imports: [RouterModule.forRoot(routes, {
+    initialNavigation: 'enabled'
+  })],
   exports: [RouterModule],
   declarations: []
 })
